make router state a singleton
