+++ src/index.js
@@ -13,6 +13,7 @@ import OnboardingModal from '@ui/OnboardingModal';
 import orientation from '@utils/orientation';
 
 import AppRouter from './AppRouter';
+import AppRoutes from './AppRoutes';
 import SentryContext from './SentryContext';
 
 const sentryEnv = Settings.APP_SENTRY_ENVIRONMENT || Settings.NODE_ENV;
@@ -32,10 +33,11 @@ const App = nest(
   withProps({ client: Client })(ApolloProvider),
   ThemeProvider,
   FontLoader,
+  AppRouter,
   ActionSheetProvider,
   SentryContext,
   OnboardingModal,
-  AppRouter,
+  AppRoutes,
 );
 
 orientation.allow(orientation.Orientation.PORTRAIT_UP);
